refactor: add index.ts files on components folders and add exports
