+++ src/App/App.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
-// import logo from './logo.svg';
 import './App.css';
-import HomePage from './routes/HomePage/HomePage';
+import HomePage from '../routes/HomePage';
 
 function App() {
   return <HomePage />;
