+++ src/components/Buttons/index.ts
@@ -1,0 +1,2 @@
+import Buttons from './Buttons';
+export default Buttons;
