+++ src/components/Modal/Modal.tsx
@@ -1,1 +1,5 @@
 import React from 'react';
+
+const Modal = () => {};
+
+export default Modal;
