+++ src/routes/HomePage/HomePage.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import Header from '../../components/Header/Header';
-import Footer from '../../components/Footer/Footer';
-import Layout from '../../components/Layout/Layout';
+import Header from '../../components/Header';
+import Footer from '../../components/Footer';
+import Layout from '../../components/Layout';
 
 const HomePage = () => {
   return (
