+++ src/routes/HomePage/index.ts
@@ -1,0 +1,2 @@
+import HomePage from './HomePage';
+export default HomePage;
